+++ sdk/storage/azure-storage-file-datalake/tests/test_file_system_async.py
@@ -106,7 +106,7 @@ class FileSystemTest(StorageTestCase):
 
         # Act
         file_system_client = self.dsc.get_file_system_client(file_system_name)
-        created = await file_system_client.create_file_system_if_not_exists()
+        created = await file_system_client.create_if_not_exists()
 
         # Assert
         self.assertTrue(created)
@@ -122,7 +122,7 @@ class FileSystemTest(StorageTestCase):
         # Act
         file_system_client = self.dsc.get_file_system_client(file_system_name)
         await file_system_client.create_file_system()
-        created = await file_system_client.create_file_system_if_not_exists()
+        created = await file_system_client.create_if_not_exists()
 
         # Assert
         self.assertIsNone(created)